chore: domains 하위로 패키지 구조 변경
