+++ src/main/java/com/ziine/domains/artist/domain/entity/ContactEntity.java
@@ -1,4 +1,4 @@
-package com.ziine.artist.domain.entity;
+package com.ziine.domains.artist.domain.entity;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.ConstraintMode;
