+++ src/main/java/com/ziine/domains/artwork/domain/repository/ArtworkRepository.java
@@ -1,6 +1,6 @@
-package com.ziine.artwork.domain.repository;
+package com.ziine.domains.artwork.domain.repository;
 
-import com.ziine.artwork.domain.entity.ArtworkEntity;
+import com.ziine.domains.artwork.domain.entity.ArtworkEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
